mttcan: Fix controller restart crash
Issue: If controller is configured in restart mode and it goes to
BUS OFF state, then kernel crash is observed.
Currently while restarting the controller, we are reading the
state from PSR register which reports the state as BUS OFF. In that
case we are informing the upper layer again about BUS OFF state.
This leads to rescheduling of can restart work which leads to crash.

Fix: No need to call bus_off again in restart, that will be handled
by bus_off irq.

There is race condition ragarding CARRIER flag, that is not handled
by CAN dev layer. Moved restart logic to driver to handle this.

Bug 200293626

Change-Id: I179775f76e4e8739f4da6047d0282521a614f4a3
Signed-off-by: Abhijit <abhijit@nvidia.com>
Reviewed-on: https://git-master.nvidia.com/r/1517956
Reviewed-by: Manoj Chourasia <mchourasia@nvidia.com>
GVS: Gerrit_Virtual_Submit
Reviewed-by: Vipin Kumar <vipink@nvidia.com>
Reviewed-by: Sandeep Trasi <strasi@nvidia.com>
Reviewed-by: mobile promotions <svcmobile_promotions@nvidia.com>
Tested-by: mobile promotions <svcmobile_promotions@nvidia.com>