video: tegra: dc: fix boot-up print message
In display terminology, term "head" should be used for physical
device that is connected to specific physical display and term
"disp" should be used for logical display device (order in which
display devices are probed, enumerated and registered with kernel).
So it is possible that head1 is exposed as disp0, /dev/fb0.

Fix the info log such a way that it aligns with above notion.

TDS-2476

Change-Id: I211d190adbfc056f996f63045db3532f1c23e395
Signed-off-by: Ujwal Patel <ujwalp@nvidia.com>
Reviewed-on: https://git-master.nvidia.com/r/1518316
Reviewed-by: Automatic_Commit_Validation_User
Reviewed-by: Shu Zhong <shuz@nvidia.com>
Reviewed-by: Aron Wong <awong@nvidia.com>
GVS: Gerrit_Virtual_Submit