dc: crc: Avoid error check for regional CRC reads
The error bit in DC_COM_RG_REGION_CRC_0 only conveys whether the HW
calculated CRCs of a region match the programmed golden CRCs for that
region. If the golden CRC registers are not programmed, this error
bit will be set every single frame end, due to the mismatch and will
generate false alarms.

Stop inferring that bit as an indicator of CRC overrun

Bug 1858958
TDS-2210

Change-Id: Ib6125696adeb9aae05f147538759bc07c6f8bd19
Signed-off-by: Jay Bhukhanwala <jbhukhanwala@nvidia.com>
Reviewed-on: https://git-master.nvidia.com/r/1522869
Reviewed-by: Automatic_Commit_Validation_User
GVS: Gerrit_Virtual_Submit
Reviewed-by: Ujwal Patel <ujwalp@nvidia.com>