soc: tegra: Don't modify rail list after init
Rather than doing an explicit sort of the dvfs_rail->dvfs list, iterate
to print the dvfs clks in a sorted manner. This prevents a list read
and a list sort to occur in parallel at which point the list read might
lead to a corrupted read. Note that there are call-sites that read the
list without the dvfs_lock held.

This path is a debugfs cat anyways so we don't care if this code is
slightly slower.

Bug 200305644

Change-Id: I4d55b40e8e37311cd5cd710af932a286d8099b75
Signed-off-by: Sai Gurrappadi <sgurrappadi@nvidia.com>
Reviewed-on: https://git-master.nvidia.com/r/1523400
Reviewed-by: Automatic_Commit_Validation_User
GVS: Gerrit_Virtual_Submit
Reviewed-by: Aleksandr Frid <afrid@nvidia.com>
Reviewed-by: Yu-Huan Hsu <yhsu@nvidia.com>