sched: Add hysteresis for capacity_aware_wake
Currently the CAPACITY_AWARE wake path tries to put a task on the
smallest capacity CPU it can find. It uses a 20% headroom to provision
future task growth but it doesn't use any hysteresis to provision for
any temporary dips in load.

This can cause unbounded thrashing if the load of a task oscillates
about the 20% headroom point. Fix this by introducing an additional 20%
guardband on the way down i.e (big->little) to provision for variablilty
in task load.

We also disable wake sync hints to let capacity_aware optimize for
throughput if it can. Not doing so can cause the wake sync path to undo
capacity_aware decisions which can cause thrashing.

Bug 1903075
Bug 200323937

Change-Id: I4df94f1a6bbd006bac47a92f1067cbe5a37dd5d9
Signed-off-by: Sai Gurrappadi <sgurrappadi@nvidia.com>
Reviewed-on: https://git-master.nvidia.com/r/1526569
Reviewed-by: Peter Boonstoppel <pboonstoppel@nvidia.com>
Reviewed-by: svccoveritychecker <svccoveritychecker@nvidia.com>
GVS: Gerrit_Virtual_Submit
Reviewed-by: Bo Yan <byan@nvidia.com>