dc: include: add scalable dc_ext IMP structs
The existing IMP structs are hardcoded with respect to
the max number of supported heads and windows across
all current SOCs. In order to fix this, this change
designs new structs in which the number of per-head
and per-window entries can be dynamically configured,
and also adds a new flip user user data type for IMP PROPOSE
requests. These structs will be used once kernel moves
to the new IMP design, and the old ones will be phased out
accordingly.

TDS-2091

Change-Id: I6c06c0c8bec8f89ac71b5268ebd731f046c81f26
Signed-off-by: Shu Zhong <shuz@nvidia.com>
Reviewed-on: https://git-master.nvidia.com/r/1526635
Reviewed-by: Ujwal Patel <ujwalp@nvidia.com>
GVS: Gerrit_Virtual_Submit
Reviewed-by: Mitch Luban <mluban@nvidia.com>