nvdisp: remove v_back_porch/v_front_porch DP quirk
Currently, the nvdisp driver adjusts the v_back_porch and
v_front_porch based on the v_ref_to_sync for DP modes. This
logic was introduced to workaround an SOR HW issue on chips
ranging from T124-T214, but does not apply to T186/T194 and
should be removed accordingly.

Bug 200332520

Change-Id: I54ad2e53ac88a395445f58e774a74f3ac1575835
Signed-off-by: Shu Zhong <shuz@nvidia.com>
Reviewed-on: https://git-master.nvidia.com/r/1527241
Reviewed-by: mobile promotions <svcmobile_promotions@nvidia.com>
Tested-by: mobile promotions <svcmobile_promotions@nvidia.com>