video: tegra: debugfs: delay on write hotplug
Many GVS scripts called by display_submit force assert/deassert
the hotplug node under debugfs.
Then they carry on execution as if the event has been processed,
which is incorrect due to the asynchronous handling of hot plug detection.
This workaround introduces a delay to make this node synchronous-like.
Without it, GVS crashes the system when FB console is enabled.
Synchronous handling needs to be implemented as a separate task.

JIRA: EVLR-1679
Bug 1858144

Change-Id: Icdacbc8e7191efd5842648a5d7dcf8e570e86a33
Signed-off-by: Tow Wang <toww@nvidia.com>
Reviewed-on: https://git-master.nvidia.com/r/1528387
Reviewed-by: mobile promotions <svcmobile_promotions@nvidia.com>
Tested-by: mobile promotions <svcmobile_promotions@nvidia.com>