r8152: use netif_receive_skb() to RX packets
r1852 driver version 2.03 used netif_receive_skb() to pass packets to
the kernel's network protocol stack. Driver 2.08 uses
napi_gro_receive(). However, this causes problems; the GRO feature
appears to fail to handle valid received packets in some cases,
leading to the remote TCP host timing out waiting for an ACK and re-
transmitting data. The retransmit appears to wake up GRO or the TCP
stack and then everything works until the problem appears again. I've
validated that this is an issue in the IP (or GRO) code and not in the
r1852 driver: a comparison of tcpdump on the server and custom debug
spew on the Tegra side indicates no packets are being lost in the
r8152 HW or driver.

Switch back to the old "receive a packet" function to solve (or hides)
this issue.

Bug 1947539

Change-Id: I05f84f9b7ad2b21e6ebf9839c5b573708ab7d9c8
Signed-off-by: Stephen Warren <swarren@nvidia.com>
Reviewed-on: https://git-master.nvidia.com/r/1529799
(cherry-picked from commit c2bff3465857b2cb4b9297f0308bb1dec8adf451 in dev-kernel-4.4)
Reviewed-on: https://git-master.nvidia.com/r/1538205
GVS: Gerrit_Virtual_Submit
Reviewed-by: Jon Mayo <jmayo@nvidia.com>