ASoC: tegra-alt: include correct header files
Kernel version checks are added in the header files. This is necessary
because k-4.9 and k-4.4 drivers are not unified at this moment and
the drivers have to include correct versions of the header files
which are getting affected by this.

Note: This change would be reverted once the unification is achieved.

Bug 200329907

Change-Id: I5972535c887c195d211971d0f77ce266f1560137
Signed-off-by: Sameer Pujar <spujar@nvidia.com>
Reviewed-on: https://git-master.nvidia.com/r/1529693
GVS: Gerrit_Virtual_Submit
Reviewed-by: Mohan Kumar D <mkumard@nvidia.com>
Reviewed-by: Ravindra Lokhande <rlokhande@nvidia.com>
Reviewed-by: Sachin Nikam <snikam@nvidia.com>
Reviewed-by: Nor Cho <ncho@nvidia.com>