soc: tegra: pmc: Do not mix two base addresses
There are two base address sharing a same point (pmc->base) in this driver:
1) Fetched in tegra_pmc_early_init()
2) Fetched in tegra_pmc_probe()

Before fetching the second one, the driver does iounmap() to the first one
in the tegra_pmc_probe(), and then reuses the same pmc->base pointer for
the second base address. However, the second base address is managed by
devres via devm_ioremap_resource(). So it will be cleaned up if the probe()
fails.

When the probe() fails with a -517 error (aka EPROBE_DEFER), the probe()
will do that iounmap() again by assuming that the first base address is
still mapped. So it would try to unmap the pointer (now it's the second
base addres) which is already being cleaned by the devres core. Then there
is a kernel warning:
 Trying to vfree() nonexistent vm area (ffff000008079000)
 ------------[ cut here ]------------
 WARNING: CPU: 3 PID: 919 at mm/vmalloc.c:1515 __vunmap+0xa8/0xc8
 Workqueue: events deferred_probe_work_func
 PC is at __vunmap+0xa8/0xc8
 LR is at __vunmap+0xa8/0xc8
 Call trace:
 [<ffff0000081936ec>] __vunmap+0xa8/0xc8
 [<ffff000008193800>] vunmap+0x20/0x2c
 [<ffff00000809b9cc>] __iounmap+0x2c/0x34
 [<ffff0000083d7d18>] tegra_pmc_probe+0x240/0xa40
 [<ffff000008516bec>] platform_drv_probe+0x50/0xbc
 [<ffff000008514948>] driver_probe_device+0x21c/0x404
 [<ffff000008514cf8>] __device_attach_driver+0x94/0x138
 [<ffff0000085127e8>] bus_for_each_drv+0x54/0x94
 [<ffff0000085145d4>] __device_attach+0xc4/0x154
 [<ffff000008514ee0>] device_initial_probe+0x10/0x18
 [<ffff000008513998>] bus_probe_device+0x90/0x98
 [<ffff000008513f94>] deferred_probe_work_func+0x98/0xe8

So this patch adds an early_base to eradicate the reuse of the pmc->base
pointer to fix the problem, and meanwhile to increase the readability.

Bug 200276674

Change-Id: I487f6052405a16ac9d6c1ef31dd497eadec00698
Signed-off-by: Nicolin Chen <nicolinc@nvidia.com>
Reviewed-on: https://git-master.nvidia.com/r/1533397
GVS: Gerrit_Virtual_Submit
Reviewed-by: svccoveritychecker <svccoveritychecker@nvidia.com>
Reviewed-by: Pritesh Raithatha <praithatha@nvidia.com>
Reviewed-by: Venkat Reddy Talla <vreddytalla@nvidia.com>
Reviewed-by: Sachin Nikam <snikam@nvidia.com>