t19x: cpuidle: update crossover function call
Crossover thresholds provide idle threshold information to
mce/mts to make better judgement as when to enter a
particular cpu power state.

The api for crossover was updated in ATF to use a scheme
where only the state information needs to be passed
instead of the actual NVG index value previously used.
This patch updates the crossover function calls to adhere
to the new api.

Change-Id: I775a14162133db09856c1782873c134a1b5574b2
Signed-off-by: Krishna Sitaraman <ksitaraman@nvidia.com>
Reviewed-on: https://git-master.nvidia.com/r/1534533
Reviewed-by: Automatic_Commit_Validation_User
Reviewed-by: svc-mobile-coverity <svc-mobile-coverity@nvidia.com>
Reviewed-by: svccoveritychecker <svccoveritychecker@nvidia.com>
Reviewed-by: Sai Gurrappadi <sgurrappadi@nvidia.com>
GVS: Gerrit_Virtual_Submit
Reviewed-by: Alexander Van Brunt <avanbrunt@nvidia.com>