dc: crc: Improve behavior of flip ID U64_MAX
When the CRC get IOCTL is issued with flip ID U64_MAX, we return the
CRCs corresponding to the most recently matched flip. But as brought
to notice, this behavior can be improved such that we return the CRCs
for the latest flip programmed through the FLIP4 IOCTL, and wait if
the CRCs for this flip are not ready.

To achieve this, for CRC get requests issued with flip ID U64_MAX:
- Retrieve the CRCs corresponding to the flip that was most recently
  queued in the flip queue
- If there are no flips pending scan out (the flip queue is empty),
  return the CRCs corresponding to the flip that was most recently
  scanned out

Bug 1858958
TDS-2506

Change-Id: Ia65edf6f538671fffa7f41d6b54ee286d6843eee
Signed-off-by: Jay Bhukhanwala <jbhukhanwala@nvidia.com>
Reviewed-on: https://git-master.nvidia.com/r/1537015
Reviewed-by: mobile promotions <svcmobile_promotions@nvidia.com>
Tested-by: mobile promotions <svcmobile_promotions@nvidia.com>