ALSA: hda: fix hda shutdown callback
During the shutdown callback the userspace is not cleanup
the process, due to which pcm streams are active even after
kernel shutdown callback. But shutdown callback in hda driver
is calling stop chip which is not right way to do as streams
are stil active, due to this iommu error is seen as stop chip
resets the adma buffer pointer during shutdown which is resulting
in iommu errors.

Bug 200331912

Change-Id: Ib751fd365430a68c2794789b288a5f6f1f20a31a
Signed-off-by: Mohan Kumar <mkumard@nvidia.com>
Reviewed-on: https://git-master.nvidia.com/r/1528815
Reviewed-on: https://git-master.nvidia.com/r/1537525
Reviewed-by: Automatic_Commit_Validation_User
Reviewed-by: svccoveritychecker <svccoveritychecker@nvidia.com>
Reviewed-by: svc-mobile-coverity <svc-mobile-coverity@nvidia.com>
GVS: Gerrit_Virtual_Submit
Reviewed-by: Sameer Pujar <spujar@nvidia.com>
Reviewed-by: Ravindra Lokhande <rlokhande@nvidia.com>