iommu: tegra-smmu: remove PLATFORM_ENABLE_IOMMU config
In K4.9 PLATFORM_ENABLE_IOMMU is not present in Kconfig and due
to that TEGRA_IOMMU_SMMU_LINEAR is getting enabled and that is
causing failures.

PLATFORM_ENABLE_IOMMU is not needed so removing it from
multiple places.

Also making TEGRA_IOMMU_SMMU_LINEAR to default n.

Bug 200277498

Change-Id: I3f316a2a2a475f7ad7e03de8ef08735b03953095
Signed-off-by: Pritesh Raithatha <praithatha@nvidia.com>
Reviewed-on: https://git-master.nvidia.com/r/1538123
Reviewed-by: svc-mobile-coverity <svc-mobile-coverity@nvidia.com>
Reviewed-by: svccoveritychecker <svccoveritychecker@nvidia.com>
Tested-by: Nicolin Chen <nicolinc@nvidia.com>
Reviewed-by: Puneet Saxena <puneets@nvidia.com>
GVS: Gerrit_Virtual_Submit
Reviewed-by: Sachin Nikam <snikam@nvidia.com>