nvdisp: remove hubparent_clk logic
pll_disphub is the HPLL source that display should always
use to generate the nvdisplayhub clk. This relationship
will now be enforced during bpmp init since it's a fixed
constraint. As such, this change removes all the logic
that we have in the nvdisp driver for configuring the
hubparent_clk.

TDS-662

Change-Id: I844b5e6735b0593a57307425e8ed96e1f8457ba6
Signed-off-by: Shu Zhong <shuz@nvidia.com>
Reviewed-on: https://git-master.nvidia.com/r/1541536
Reviewed-by: Automatic_Commit_Validation_User
Reviewed-by: Ujwal Patel <ujwalp@nvidia.com>
Reviewed-by: Arun Swain <arswain@nvidia.com>
Reviewed-by: svc-mobile-coverity <svc-mobile-coverity@nvidia.com>
Reviewed-by: svccoveritychecker <svccoveritychecker@nvidia.com>
GVS: Gerrit_Virtual_Submit
Reviewed-by: Mitch Luban <mluban@nvidia.com>