tegra: camera: rtcpu: add memory areas for tests
Map and pass memory areas to the RTCPU functional tests.

There are 8 memory areas, accessible from debugfs:

/d/camrtc/test/mem[0-7]

The memory areas get mapped to the RTCPU address space. The
nvidia,mem-map devicetree property can be used to optionally map the
memory to VI and ISP address spaces.

The sizes and base addresses of the memory areas are passed to the
RTCPU in a debug message. After executing the test, the sizes of the
memory areas are updated according to the sizes in return message.

Jira CAMC-446

Change-Id: Ibf71d8043920aebc9c54a5588d0671585b198d0c
Signed-off-by: Pekka Pessi <ppessi@nvidia.com>
Reviewed-on: https://git-master.nvidia.com/r/1545114
Reviewed-by: Automatic_Commit_Validation_User
Reviewed-by: svc-mobile-coverity <svc-mobile-coverity@nvidia.com>
Reviewed-by: svccoveritychecker <svccoveritychecker@nvidia.com>
Reviewed-by: Semi Malinen <smalinen@nvidia.com>
GVS: Gerrit_Virtual_Submit
Reviewed-by: Jihoon Bang <jbang@nvidia.com>