video: tegra: host: Always use VM threshold regs
The WAR to detect which threshold register range to use is flaky. Older
simulator models no longer need to be supported plus this is the
configuration we'll use on silicon so just yank the WAR.

Using the wrong threshold register range ends up causing an interrupt
storm.

Bug 1977648

Change-Id: I7c9c80869329d74702eacb4a9099f36c04a5bf6a
Signed-off-by: Sai Gurrappadi <sgurrappadi@nvidia.com>
Reviewed-on: https://git-master.nvidia.com/r/1545481
Reviewed-by: Automatic_Commit_Validation_User
Reviewed-by: Arto Merilainen <amerilainen@nvidia.com>
Reviewed-by: Mikko Perttunen <mperttunen@nvidia.com>
Reviewed-by: svc-mobile-coverity <svc-mobile-coverity@nvidia.com>
Reviewed-by: svccoveritychecker <svccoveritychecker@nvidia.com>
Reviewed-by: Seshendra Gadagottu <sgadagottu@nvidia.com>
Tested-by: Seshendra Gadagottu <sgadagottu@nvidia.com>
GVS: Gerrit_Virtual_Submit
Reviewed-by: Bo Yan <byan@nvidia.com>