tegra: dc: Maintain per ext count of users
We don't have a way to track if there is an active userspace process
that has opened the file descriptor corresponding to a tegra ext
device. This can be useful to know if certain cleanups (for example,
resetting CRC reference counts for the head) are safe to do.

- Add and maintain a per ext count of users that have opened the file
  descriptor of the tegra ext device file

Bug 1858958
TDS-2612

Change-Id: I3dbd8bc10b51d19f923796f967db9088c7b5ed41
Signed-off-by: Jay Bhukhanwala <jbhukhanwala@nvidia.com>
Reviewed-on: https://git-master.nvidia.com/r/1545974
Reviewed-by: Shu Zhong <shuz@nvidia.com>
Reviewed-by: svc-mobile-coverity <svc-mobile-coverity@nvidia.com>
GVS: Gerrit_Virtual_Submit
Reviewed-by: svccoveritychecker <svccoveritychecker@nvidia.com>
Reviewed-by: Automatic_Commit_Validation_User
Reviewed-by: Ujwal Patel <ujwalp@nvidia.com>