media: tegra: camera: capture: Fix unpin lists
The unpin list allocated for the capture channel was
being allocated and copied to incorrectly, leading to
unpin data being lost and the resources never unpinned.

This changes the capture channel's unpin list to store
an array of unpin lists that are alloc/pinned/assigned to
each capture when the request is submitted, then unpinned
and freed when the capture completes.

Change-Id: I3151fb106b080c72da1908d0665767ae4a7ae481
Signed-off-by: Ian Stewart <istewart@nvidia.com>
Reviewed-on: https://git-master.nvidia.com/r/1549412
Reviewed-by: svc-mobile-coverity <svc-mobile-coverity@nvidia.com>
GVS: Gerrit_Virtual_Submit
Reviewed-by: svccoveritychecker <svccoveritychecker@nvidia.com>
Reviewed-by: Jihoon Bang <jbang@nvidia.com>