scripts: Add an 'rfr' program
Add an RFR program to generate RFR emails. Since there are many emails
sent with either no information or lacking information this script
will help make RFR emails more useful to developers interested in
conducting reviews.

The help is as follows:

  $ ./rfr --help
  usage: rfr [-h] [-V] [-m MSG] Commit-IDs [Commit-IDs ...]

  RFR formatting tool

  positional arguments:
    Commit-IDs         Git or gerrit commits to describe. Can be either a git or
                       gerrit commit ID. If the ID starts with a 'I' then it
                       will be treated as a gerrit ID. If the commit ID looks
                       like a gerrit URL then it is treated as a gerrit URL.
                       Otherwise it's treated as a git commit ID.

  optional arguments:
    -h, --help         show this help message and exit
    -V, --version      print the program version
    -m MSG, --msg MSG  Custom message to add to the RFR email

  This program will format commit messages into something that can be sent to
  the nvgpu mailing list for review

The Commit IDs can be a git commit ID, gerrit change ID, or a gerrit URL. For
example:

  # Git commit IDs (abbreviated are ok)
  $ ./rfr 89e2470807
  $ ./rfr 89e2470807571ef6dc545ece8f5f7e43e27b0c12

  # Gerrit Change-Id (also can be abbreviated)
  $ ./rfr I0524ae1982fbce9eb4da976e72102e08e2dcd017
  $ ./rfr I0524ae198

  # And lastly, several different URL formats:
  https://git-master.nvidia.com/r/#/c/1553782/2
  https://git-master.nvidia.com/r/#/c/1553782/
  https://git-master.nvidia.com/r/1553782
  https://git-master/r/1553782

The only currently known bug is that if yo have multiple versions of a
commit that share the same gerrit Change-ID then this script will only
take the first one. For example if there's a bunch of cherry-picks then
which commit is taken is undefined.

It seems like generally the order that gerrit returns changes puts open
changes first, but this is not tested nor should this be relied upon. At
some point supporting some mechanism to filter will be necessary.

Change-Id: I0524ae1982fbce9eb4da976e72102e08e2dcd017
Signed-off-by: Alex Waterman <alexw@nvidia.com>
Reviewed-on: https://git-master.nvidia.com/r/1552173
Reviewed-by: Terje Bergstrom <tbergstrom@nvidia.com>