arm64: tegra_defconfig: Enable TEGRA_SOCTHERM
The TEGRA_POWERGATE is default =y in Kconfig. And its powergate-t21x.c
calls a function in soctherm drivers. Once the powergate-t21x is enabled
by correcting all T21x_SOC to T210_SOC, it will have a compiling error
as the soctherm driver is not compiled at all with tegra_defconfig:
  drivers/platform/tegra/powergate/powergate-t21x.c:1299:
    undefined reference to `tegra_soctherm_gpu_tsens_invalidate'

So this patch enables TEGRA_SOCTHERM in the tegra_defconfig to fix the
compiling error.

Bug 200340638

Change-Id: Ida5b4e1c623a7fea7b74f018388faedf5ab3f079
Signed-off-by: Nicolin Chen <nicolinc@nvidia.com>
Reviewed-on: https://git-master.nvidia.com/r/1553102
Reviewed-by: Automatic_Commit_Validation_User
GVS: Gerrit_Virtual_Submit
Reviewed-by: Bharat Nihalani <bnihalani@nvidia.com>