video: tegra: host: Support flcn fw reloading
Currently the flcn driver loads the firmware just once on device power up.
It is userful though to have the ability to reload the firmware at any
point for testing / development.

To that end, add a sysfs node that force idle's the device and reinits the
fw so that on the subsequent unidle, the fw is reloaded.

JIRA HOSTX-241

Change-Id: I61b674b7351df48547121a7a4d3e169be3d6d586
Signed-off-by: Sai Gurrappadi <sgurrappadi@nvidia.com>
Reviewed-on: https://git-master.nvidia.com/r/1555779
Reviewed-by: mobile promotions <svcmobile_promotions@nvidia.com>
Tested-by: mobile promotions <svcmobile_promotions@nvidia.com>