media:platform:tegra: Use cil_settletime from DT
At the V4L2 level, we were ignoring the sensor device tree node's
cil_settletime property, and using some default (DEFAULT_THS_SETTLE) which
does not work for all sensors.

The cil_settletime sensor property is parsed from the sensor's device tree
node and stored in a struct camera_common_data.

Weirdly, struct tegra_csi_channel has a pointer to struct
camera_common_data (s_data) that's completely unused.  This patch assumes
that the intent of this pointer is for the tegra_csi_channel to be able to
get at the sensor properties.  After parsing the sensor properties, we
traverse the device tree to get at the csi node on which this sensor is
attached, and connect the two via tegra_csi_channel's s_data pointer.

Also, if cil_settletime is 0, autocalculate the value rather than using a
hardcoded default.

Bug 1982959
Bug 1885366
Bug 1981695

Change-Id: Ic340e984b20ac3c81d76ad5771c4df077c00f286
Signed-off-by: Josh Kergan <jkergan@nvidia.com>
Reviewed-on: https://git-master.nvidia.com/r/1554934
Reviewed-by: mobile promotions <svcmobile_promotions@nvidia.com>
Tested-by: mobile promotions <svcmobile_promotions@nvidia.com>