dc: crc: Fix RG/COMP CRC read programming sequence
There is a subtle difference in the current code and the programming
sequence in the manual. The correct sequence requires reading the
CRC_B register every frame end interrupt when the valid bit is set,
to avoid setting the error in the successive frame. Also, there is
lack of information printed on the UART Log in error case

- Fix the programming sequence of RG and COMP CRC reads
- Print HW frame number, status and control register values to the
  error log

Bug 1987050
TDS-2706

Change-Id: I612664734464f711ba335add27b82623be34febb
Signed-off-by: Jay Bhukhanwala <jbhukhanwala@nvidia.com>
Reviewed-on: https://git-master.nvidia.com/r/1557434
Reviewed-by: svccoveritychecker <svccoveritychecker@nvidia.com>
GVS: Gerrit_Virtual_Submit
Reviewed-by: Ujwal Patel <ujwalp@nvidia.com>